requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pandas==2.1.3
numpy==1.25.2
websockets==11.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
celery==5.3.4
pytest==7.4.3
aiofiles==0.24.0
python-dotenv==1.0.0
requests==2.32.3
lxml==4.9.4
yfinance==0.2.43
sentry-sdk==1.39.1
pydantic==1.10.13